requirements.txt
torch>=2.0
torchvision
numpy
scikit-learn
tensorflow-datasets
matplotlib
faiss-cpu
scipy
